chore: React/CSS 분리 및 폴더 구조 변경
