+++ src/components/HomePage/HomeItems/CalendarSection/Calendar.jsx
@@ -1,85 +1,15 @@
 import React, { useEffect } from 'react';
-import styled from 'styled-components';
 import PropTypes from 'prop-types';
 
-const CalendarContainer = styled.div`
-  display: grid;
-  grid-template-columns: repeat(7, 0fr);
-  grid-column-gap: 0px;
-  justify-content: center;
-`;
-
-const Day = styled.div`
-  display: flex;
-  font-weight: bold;
-  width: 3.5vw;
-  height: 4.3vw;
-  padding-left: 5px;
-  padding-top: 5px;
-  z-index: 1;
-  background-color: transparent;
-  color: inherit;
-  opacity: ${({ isOutsideMonth }) => (isOutsideMonth ? 0.3 : 1)};
-  position: relative;
-  &.frame {
-    height: 2vw;
-    color: ${({ day }) => (day === 'SUN' ? 'red' : 'inherit')};
-  }
-`;
-
-const Circle = styled.div`
-  height: 1.5vw;
-  width: 1.5vw;
-  background-color: #c6ef5e; /* 동그라미의 배경색 설정, 필요에 따라 변경 가능 */
-  border-radius: 50%;
-  position: absolute; /* 동그라미의 위치를 상대적으로 설정한 부모에 대해 설정 */
-  margin-top: -4px;
-  margin-left: 9px;
-  justify-content: center;
-  align-items: center;
-  opacity: 0.5;
-  z-index: 2;
-  transform: translateX(-50%); /* 가운데 정렬을 위해 50%만큼 이동 */
-  display: ${({ isSelected }) =>
-    isSelected
-      ? 'flex'
-      : 'none'}; /* 선택된 날에만 동그라미를 표시하도록 설정 */
-`;
-
-const Month = styled.h2`
-  text-align: right;
-  padding-right: 2vw;
-  margin-bottom: 1vw;
-`;
-
-const DayFont = styled.p`
-  width: 25%;
-  text-align: center;
-  position: absolute;
-  z-index: 3;
-  margin-left: 0px;
-  margin-top: 0px;
-`;
-
-const ChartBox = styled.div`
-  display: flex;
-  flex-direction: column;
-  margin-top: 50%;
-  height: 50%;
-  width: 100%;
-  align-items: baseline;
-`;
-
-const Chart = styled.div`
-  display: flex;
-  flex-direction: column;
-  height: 18%;
-  width: 100%;
-  margin-bottom: 2px;
-  padding-left: 8%;
-  background-color: ${(props) =>
-    props.isSelected ? '#90C20D' : 'transparent'};
-`;
+import {
+  CalendarContainer,
+  Day,
+  Circle,
+  Month,
+  DayFont,
+  ChartBox,
+  Chart,
+} from './CalendarSectionItemCss';
 
 function ChartHandler(projects) {
   projects.forEach((project) => {
