+++ src/components/HomePage/HomeItems/CalendarSection/CalendarSection.jsx
@@ -1,19 +1,9 @@
-import styled from 'styled-components';
-import { Title } from '../Home';
 import Calendar from './Calendar';
 import { useNavigate } from 'react-router-dom';
 import PropTypes from 'prop-types';
 
-const CalendarBox = styled.div`
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  background-color: #ecf1e0;
-  width: 29vw;
-  height: 29.5vw;
-  border: 3px solid #b9d967;
-  border-radius: 30px 30px 30px 30px;
-`;
+import { CalendarBox } from './CalendarSectionItemCss';
+import { Title } from '../HomeItemCss';
 
 export default function CalendarSection({
   calendarIdx,
