+++ src/components/HomePage/HomeItems/HomeItemCss.jsx
@@ -1,0 +1,25 @@
+import styled from 'styled-components';
+
+export const Title = styled.h3`
+  cursor: pointer;
+  font-size: 20px;
+  margin-bottom: 1.5vh;
+`;
+
+export const Top = styled.div`
+  display: flex;
+  justify-content: center;
+`;
+
+export const Left = styled.div`
+  margin-left: 19.5vw;
+`;
+
+export const Right = styled.div`
+  margin-left: 3vw;
+  margin-bottom: 3vw;
+`;
+
+export const Bottom = styled.div`
+  display: flex;
+`;
